Clean up/Organization - RERUN install.sh
Moved 'Move' class and children to 'auto_speed_move.py'
Moved 'ResultsWrapper' and 'AttemptWrapper' to 'auto_speed_wrappers.py'
Moved generic functions to 'auto_speed_funcs.py'
Updated 'instal.sh' to link all files ending with .py
Cleaned up results_dir code to match code style
Updated/added comments for clarity
Updated README for clarity
Added more items to roadmap
Updated README example configuration to work out-of-the box
Commented out example config to use internal defaults by default if copied
Updated console output for clarity
